requirements.txt
streamlit==1.50.0
groq==0.31.1
python-dotenv==1.1.1
PyYAML==6.0.3
pandas==2.3.2
numpy==2.3.3
requests==2.32.5
Flask==3.1.2
